requirements.txt
Flask==3.0.0
werkzeug==3.1.0
openai==1.52.2
python-dotenv==1.0.1
flask-cors==5.0.0
gunicorn==21.2.0
pyaudio==0.2.14
